+++ src/drawable.rs
@@ -22,7 +22,6 @@ use cairo::Context;
 use shakmaty::Square;
 
 use ground::{EventContext, GroundMsg};
-use util::float;
 
 /// Shape colors.
 #[derive(Copy, Clone, Eq, PartialEq, Hash, Debug)]
@@ -170,10 +169,10 @@ impl DrawShape {
             DrawBrush::Yellow => cr.set_source_rgba(0.90, 0.94, 0.0, opacity),
         }
 
-        let orig_x = 0.5 + float(self.orig.file());
-        let orig_y = 7.5 - float(self.orig.rank());
-        let dest_x = 0.5 + float(self.dest.file());
-        let dest_y = 7.5 - float(self.dest.rank());
+        let orig_x = 0.5 + f64::from(self.orig.file());
+        let orig_y = 7.5 - f64::from(self.orig.rank());
+        let dest_x = 0.5 + f64::from(self.dest.file());
+        let dest_y = 7.5 - f64::from(self.dest.rank());
 
         if self.is_circle() {
             // draw circle
